Implemented deleteOrder() handling not existing orders

+++ com.giulia.giamberini.book.orders/src/main/java/com/giulia/giamberini/book/orders/controller/OrderController.java
@@ -30,6 +30,11 @@ public class OrderController {
 	}
 
 	public void deleteOrder(Order orderToRemove) {
+		if (orderRepository.findByID(orderToRemove.getID()) == null) {
+			bookOrdersView.showErrorNoOrderFound("The order with id " + orderToRemove.getID() + " is not found",
+					orderToRemove);
+			return;
+		}
 		orderRepository.delete(orderToRemove.getID());
 		bookOrdersView.orderRemoved(orderToRemove);
 	}
